fix(hydration): resolve browser extension compatibility with client-side theme provider and suppressHydrationWarning
